+++ pkg/pipelines/config.go
@@ -1,23 +1,13 @@
 package pipelines
 
-type Mode int
-
-const (
-	ModeRun Mode = iota
-	ModeDryRun
-)
-
 type Config struct {
-	CacheDir         string `json:"cacheDir"`
-	DaggerDebugImage string `json:"debugImage"`
-	DaggerExec       string `json:"daggerExec"`
-	Image            Image  `json:"image"`
+	CacheDir   string `json:"cacheDir"`
+	DaggerExec string `json:"daggerExec"`
+	Image      Image  `json:"image"`
 }
 
 func SetDefaults(c *Config) {
 	c.CacheDir = valueOrDefault(c.CacheDir, "./wfe-cache")
-	c.DaggerDebugImage = valueOrDefault(c.DaggerDebugImage, "alpine:latest")
-	c.DaggerExec = valueOrDefault(c.DaggerExec, "dagger")
 	c.Image.BuildDir = valueOrDefault(c.Image.BuildDir, ".")
 	c.Image.BuildDockerfile = valueOrDefault(c.Image.BuildDockerfile, "Dockerfile")
 	c.Image.BuildPlatform = valueOrDefault(c.Image.BuildPlatform, "")