+++ pkg/shell/podman.go
@@ -4,15 +4,15 @@ import (
 	"io"
 )
 
-type podmanCmd struct {
+type dockerCLICmd struct {
 	Executable
 	InitCmd func() *Executable
 }
 
-// Version outputs the version of the podman CLI
+// Version outputs the version of the CLI
 //
-// shell: `podman version`
-func (p *podmanCmd) Version() *Command {
+// shell: `[docker|podman] version`
+func (p *dockerCLICmd) Version() *Command {
 	cmd := p.InitCmd().WithArgs("version")
 	return &Command{
 		RunFunc: func() error {
@@ -22,12 +22,35 @@ func (p *podmanCmd) Version() *Command {
 	}
 }
 
-// PodmanComand with custom stdout and stderr
-func PodmanComand(stdout io.Writer, stderr io.Writer) *podmanCmd {
-	return &podmanCmd{
+// Info tests the connection to the container runtime daemon
+//
+// shell: `[docker|podman] info`
+func (p *dockerCLICmd) Info() *Command {
+	cmd := p.InitCmd().WithArgs("info")
+	return &Command{
+		RunFunc: func() error {
+			return cmd.Run()
+		},
+		DebugInfo: cmd.String(),
+	}
+}
+
+// PodmanCommand with custom stdout and stderr
+func PodmanCommand(stdout io.Writer, stderr io.Writer) *dockerCLICmd {
+	return &dockerCLICmd{
 		Executable: Executable{},
 		InitCmd: func() *Executable {
 			return NewExecutable("podman").WithOutput(stdout).WithStderr(stderr)
+		},
+	}
+}
+
+// DockerCommand with custom stdout and stderr
+func DockerCommand(stdout io.Writer, stderr io.Writer) *dockerCLICmd {
+	return &dockerCLICmd{
+		Executable: Executable{},
+		InitCmd: func() *Executable {
+			return NewExecutable("docker").WithOutput(stdout).WithStderr(stderr)
 		},
 	}
 }